UI for registration screen
