+++ app/src/main/java/rustam/urazov/budgetoffamily/screen/signup/SignUpScreenFactory.kt
@@ -1,0 +1,11 @@
+package rustam.urazov.budgetoffamily.screen.signup
+
+import android.content.Context
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+
+class SignUpScreenFactory(context: Context) : ViewModelProvider.Factory {
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+        TODO("Not yet implemented")
+    }
+}
