+++ app/src/main/java/rustam/urazov/budgetoffamily/screen/signup/SignUpScreenViewModel.kt
@@ -1,0 +1,8 @@
+package rustam.urazov.budgetoffamily.screen.signup
+
+import androidx.lifecycle.ViewModel
+
+class SignUpScreenViewModel : ViewModel() {
+
+
+}
