+++ app/src/main/res/layout/fragment_sign_up.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".screen.signin.SignInScreen">
+    tools:context=".screen.signup.SignUpFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
@@ -22,17 +22,34 @@
             app:layout_constraintRight_toRightOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
 
-        <TextView
-            android:id="@+id/tvAppName"
-            android:layout_width="wrap_content"
+        <EditText
+            android:id="@+id/etFirstName"
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_marginTop="36dp"
-            android:text="@string/app_name"
+            android:layout_marginTop="12dp"
+            android:layout_marginBottom="12dp"
+            android:autofillHints="name"
+            android:backgroundTint="@color/dark"
+            android:hint="@string/first_name"
+            android:inputType="text"
             android:textColor="@color/dark"
-            android:textSize="30sp"
-            app:layout_constraintLeft_toLeftOf="parent"
-            app:layout_constraintRight_toRightOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/ivLogo" />
+            app:layout_constraintBottom_toTopOf="@id/etLastName"
+            app:layout_constraintTop_toBottomOf="@id/ivLogo"
+            app:layout_constraintVertical_bias="1" />
+
+        <EditText
+            android:id="@+id/etLastName"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginBottom="12dp"
+            android:autofillHints="name"
+            android:backgroundTint="@color/dark"
+            android:hint="@string/last_name"
+            android:inputType="text"
+            android:textColor="@color/dark"
+            app:layout_constraintBottom_toTopOf="@id/etEmail"
+            app:layout_constraintTop_toBottomOf="@id/etFirstName"
+            app:layout_constraintVertical_bias="1" />
 
         <EditText
             android:id="@+id/etEmail"
@@ -45,17 +62,31 @@
             android:inputType="textEmailAddress"
             android:textColor="@color/dark"
             app:layout_constraintBottom_toTopOf="@id/etPassword"
-            app:layout_constraintTop_toBottomOf="@id/tvAppName"
+            app:layout_constraintTop_toBottomOf="@id/etLastName"
             app:layout_constraintVertical_bias="1" />
 
         <EditText
             android:id="@+id/etPassword"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_marginBottom="72dp"
+            android:layout_marginBottom="12dp"
             android:autofillHints="password"
             android:backgroundTint="@color/dark"
             android:hint="@string/password"
+            android:inputType="textPassword"
+            android:textColor="@color/dark"
+            app:layout_constraintBottom_toTopOf="@id/etPasswordAgain"
+            app:layout_constraintTop_toBottomOf="@id/etEmail"
+            app:layout_constraintVertical_bias="1" />
+
+        <EditText
+            android:id="@+id/etPasswordAgain"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginBottom="72dp"
+            android:autofillHints="password"
+            android:backgroundTint="@color/dark"
+            android:hint="@string/password_again"
             android:inputType="textPassword"
             android:textColor="@color/dark"
             app:layout_constraintBottom_toTopOf="@id/bSignIn" />
@@ -66,7 +97,7 @@
             android:layout_height="wrap_content"
             android:layout_marginBottom="18dp"
             android:background="@drawable/button_corners_12dp"
-            android:text="@string/signIn"
+            android:text="@string/signUp"
             android:textColor="@color/light"
             android:textSize="21sp"
             app:backgroundTint="@color/yellow"
@@ -77,10 +108,11 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:background="@drawable/button_corners_12dp"
-            android:text="@string/registration"
+            android:text="@string/have_account"
             android:textColor="@color/light"
             android:textSize="21sp"
             app:backgroundTint="@color/orange"
             app:layout_constraintBottom_toBottomOf="parent" />
     </androidx.constraintlayout.widget.ConstraintLayout>
+
 </FrameLayout>
