+++ data/src/main/java/rustam/urazov/budgetoffamily/network/models/auth/AuthBody.kt
@@ -1,0 +1,3 @@
+package rustam.urazov.budgetoffamily.network.models.auth
+
+data class AuthBody(val email: String, val password: String)
