+++ data/src/main/java/rustam/urazov/budgetoffamily/network/models/auth/AuthResponse.kt
@@ -1,0 +1,5 @@
+package rustam.urazov.budgetoffamily.network.models.auth
+
+import rustam.urazov.budgetoffamily.models.Token
+
+data class AuthResponse(val token: Token)
