+++ data/src/main/java/rustam/urazov/budgetoffamily/network/models/auth/TokenNetwork.kt
@@ -1,0 +1,3 @@
+package rustam.urazov.budgetoffamily.network.models.auth
+
+data class TokenNetwork(val accessToken: String)
