+++ data/src/main/java/rustam/urazov/budgetoffamily/repositories/TokenRepositoryImpl.kt
@@ -1,0 +1,16 @@
+package rustam.urazov.budgetoffamily.repositories
+
+import rustam.urazov.budgetoffamily.storage.TokenStorageService
+import rustam.urazov.budgetoffamily.storage.models.TokenStorage
+import rustam.urazov.budgetoffamily.models.Token
+
+class TokenRepositoryImpl(private val tokenStorageService: TokenStorageService): TokenRepository {
+
+    override fun saveToken(token: Token) {
+        val accessToken = TokenStorage(token.accessToken)
+
+        tokenStorageService.save(accessToken)
+    }
+
+    override fun getToken() = Token(tokenStorageService.get().token)
+}
