+++ data/src/main/java/rustam/urazov/budgetoffamily/storage/TokenStorageService.kt
@@ -1,0 +1,10 @@
+package rustam.urazov.budgetoffamily.storage
+
+import rustam.urazov.budgetoffamily.storage.models.TokenStorage
+
+interface TokenStorageService {
+
+    fun save(tokenStorage: TokenStorage)
+
+    fun get(): TokenStorage
+}
