+++ data/src/main/java/rustam/urazov/budgetoffamily/storage/models/TokenStorage.kt
@@ -1,0 +1,3 @@
+package rustam.urazov.budgetoffamily.storage.models
+
+data class TokenStorage(val token: String)
