--- data/src/main/java/rustam/urazov/data/network/API.kt
@@ -1,18 +1,0 @@
-package rustam.urazov.data.network
-
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-class API {
-    companion object {
-        const val BASE_URL = "based"
-        val mInstance = API()
-    }
-
-    private val retrofit = Retrofit.Builder()
-        .baseUrl(BASE_URL)
-        .addConverterFactory(GsonConverterFactory.create())
-        .build()
-
-    val service = retrofit.create(Service::class.java)
-}
