--- data/src/main/java/rustam/urazov/data/network/models/auth/AuthBody.kt
@@ -1,3 +1,0 @@
-package rustam.urazov.data.network.models.auth
-
-data class AuthBody(val email: String, val password: String)
