--- data/src/main/java/rustam/urazov/data/network/models/auth/AuthResponse.kt
@@ -1,5 +1,0 @@
-package rustam.urazov.data.network.models.auth
-
-import rustam.urazov.domain.models.Token
-
-data class AuthResponse(val token: Token)
