--- data/src/main/java/rustam/urazov/data/network/models/auth/TokenNetwork.kt
@@ -1,3 +1,0 @@
-package rustam.urazov.data.network.models.auth
-
-data class TokenNetwork(val accessToken: String)
