--- data/src/main/java/rustam/urazov/data/storage/TokenStorageService.kt
@@ -1,10 +1,0 @@
-package rustam.urazov.data.storage
-
-import rustam.urazov.data.storage.models.TokenStorage
-
-interface TokenStorageService {
-
-    fun save(tokenStorage: TokenStorage)
-
-    fun get(): TokenStorage
-}
