--- data/src/main/java/rustam/urazov/data/storage/TokenStorageServiceImpl.kt
@@ -1,25 +1,0 @@
-package rustam.urazov.data.storage
-
-import android.content.Context
-import rustam.urazov.data.storage.models.TokenStorage
-
-class TokenStorageServiceImpl(private val context: Context) : TokenStorageService {
-
-    companion object {
-
-        const val SETTINGS = "settings"
-        const val ACCESS_TOKEN = "access_token"
-    }
-
-    private val sharedPreferences = context.getSharedPreferences(SETTINGS, Context.MODE_PRIVATE)
-
-    override fun save(tokenStorage: TokenStorage) {
-
-        sharedPreferences.edit().putString(ACCESS_TOKEN, tokenStorage.token).apply()
-    }
-
-    override fun get(): TokenStorage {
-
-        return TokenStorage(sharedPreferences.getString(ACCESS_TOKEN, "") ?: "")
-    }
-}
