--- data/src/main/java/rustam/urazov/data/storage/models/TokenStorage.kt
@@ -1,3 +1,0 @@
-package rustam.urazov.data.storage.models
-
-data class TokenStorage(val token: String)
