+++ domain/src/main/java/rustam/urazov/budgetoffamily/models/Token.kt
@@ -1,0 +1,3 @@
+package rustam.urazov.budgetoffamily.models
+
+data class Token(val accessToken: String)
