+++ domain/src/main/java/rustam/urazov/budgetoffamily/models/UserAuthData.kt
@@ -1,0 +1,3 @@
+package rustam.urazov.budgetoffamily.models
+
+data class UserAuthData(val email: String, val password: String)
