+++ domain/src/main/java/rustam/urazov/budgetoffamily/models/error/ErrorResponse.kt
@@ -1,0 +1,3 @@
+package rustam.urazov.budgetoffamily.models.error
+
+data class ErrorResponse(val message: String)
