+++ domain/src/main/java/rustam/urazov/budgetoffamily/repositories/TokenRepository.kt
@@ -1,0 +1,10 @@
+package rustam.urazov.budgetoffamily.repositories
+
+import rustam.urazov.budgetoffamily.models.Token
+
+interface TokenRepository {
+
+    fun saveToken(token: Token)
+
+    fun getToken(): Token
+}
