+++ domain/src/main/java/rustam/urazov/budgetoffamily/repositories/UserAuthorizationRepository.kt
@@ -1,0 +1,8 @@
+package rustam.urazov.budgetoffamily.repositories
+
+import rustam.urazov.budgetoffamily.ResultWrapper
+import rustam.urazov.budgetoffamily.models.UserAuthData
+
+interface UserAuthorizationRepository {
+    suspend fun authorize(userAuthData: UserAuthData): ResultWrapper<Any>
+}
