+++ domain/src/main/java/rustam/urazov/budgetoffamily/usecases/SaveTokenUseCase.kt
@@ -1,0 +1,11 @@
+package rustam.urazov.budgetoffamily.usecases
+
+import rustam.urazov.budgetoffamily.models.Token
+import rustam.urazov.budgetoffamily.repositories.TokenRepository
+
+class SaveTokenUseCase(private val tokenRepository: TokenRepository) {
+
+    fun execute(token: Token) {
+        tokenRepository.saveToken(token)
+    }
+}
