+++ domain/src/main/java/rustam/urazov/budgetoffamily/usecases/UserAuthorizationUseCase.kt
@@ -1,0 +1,9 @@
+package rustam.urazov.budgetoffamily.usecases
+
+import rustam.urazov.budgetoffamily.models.UserAuthData
+import rustam.urazov.budgetoffamily.repositories.UserAuthorizationRepository
+
+class UserAuthorizationUseCase(private val userAuthorizationRepository: UserAuthorizationRepository) {
+
+    suspend fun execute(userAuthData: UserAuthData) = userAuthorizationRepository.authorize(userAuthData)
+}
