--- domain/src/main/java/rustam/urazov/domain/ResultWrapper.kt
@@ -1,10 +1,0 @@
-package rustam.urazov.domain
-
-import rustam.urazov.domain.models.error.ErrorResponse
-
-sealed class ResultWrapper<out T> {
-    data class Success<out T>(val value: T): ResultWrapper<T>()
-    data class Error(val code: Int? = null, val error: ErrorResponse? = null): ResultWrapper<Nothing>()
-
-    object NetworkError: ResultWrapper<Nothing>()
-}
