--- domain/src/main/java/rustam/urazov/domain/models/Token.kt
@@ -1,3 +1,0 @@
-package rustam.urazov.domain.models
-
-data class Token(val accessToken: String)
