--- domain/src/main/java/rustam/urazov/domain/models/UserAuthData.kt
@@ -1,3 +1,0 @@
-package rustam.urazov.domain.models
-
-data class UserAuthData(val email: String, val password: String)
