--- domain/src/main/java/rustam/urazov/domain/models/error/ErrorResponse.kt
@@ -1,3 +1,0 @@
-package rustam.urazov.domain.models.error
-
-data class ErrorResponse(val message: String)
