--- domain/src/main/java/rustam/urazov/domain/repositories/TokenRepository.kt
@@ -1,10 +1,0 @@
-package rustam.urazov.domain.repositories
-
-import rustam.urazov.domain.models.Token
-
-interface TokenRepository {
-
-    fun saveToken(token: Token)
-
-    fun getToken(): Token
-}
