--- domain/src/main/java/rustam/urazov/domain/usecases/SaveTokenUseCase.kt
@@ -1,11 +1,0 @@
-package rustam.urazov.domain.usecases
-
-import rustam.urazov.domain.models.Token
-import rustam.urazov.domain.repositories.TokenRepository
-
-class SaveTokenUseCase(private val tokenRepository: TokenRepository) {
-
-    fun execute(token: Token) {
-        tokenRepository.saveToken(token)
-    }
-}
